refactor(future-cli): split out evaluation of source from outputData generation
